ready for deploy

+++ handlers/name.py
@@ -4,7 +4,7 @@ from aiogram.dispatcher import FSMContext
 from constants.message import MESSAGES
 from modules.keyboards.location import ikb_menu
 from modules.quiz import Quiz
-from core.dp import dp
+from modules.dp import dp
 
 
 @dp.message_handler(state=Quiz.name)
@@ -12,12 +12,14 @@ async def send_message(message: types.Message, state=FSMContext):
     async with state.proxy() as quiz_responses:
         name = message.text
         quiz_responses["name"] = name
-    await message.delete()
-    await message.reply_to_message.delete()
+    if message.reply_to_message:
+        await message.reply_to_message.delete()
     await message.answer(
-        MESSAGES.name.substitute(name=name),
+        MESSAGES.where_are_you.substitute(name=name),
         parse_mode='HTML',
         reply_markup=ikb_menu
     )
+    await message.delete()
+
     await Quiz.next()
